+++ src/value.cpp
@@ -1028,7 +1028,7 @@ namespace Exiv2 {
         tms.tm_mday = date_.day;
         tms.tm_mon = date_.month - 1;
         tms.tm_year = date_.year - 1900;
-        long l = static_cast<long>(std::mktime(&tms));
+        long l = std::mktime(&tms);
         ok_ = (l != -1);
         return l;
     }
